+++ frontend/src/components/SemesterBlock.tsx
@@ -5,13 +5,11 @@ import { AddBlock } from "./ClassBlocks/AddBlock";
 import { AddClass, ClassList, EmptyBlock } from ".";
 import {
   DNDScheduleTerm,
-  ScheduleCourse,
   CourseWarning,
   DNDScheduleCourse,
-  IWarning,
-  Status,
-  SeasonWord,
+  IWarning
 } from "../models/types";
+import {ScheduleCourse, Status, SeasonWord} from "graduate-common";
 import styled from "styled-components";
 import { AppState } from "../state/reducers/state";
 import { connect } from "react-redux";
