+++ frontend/src/plans/index.ts
@@ -2,7 +2,7 @@ import { bioChemPlans } from "./plans-2018-biochem";
 import { mathPlans } from "./plans-2018-math";
 import { csPlans } from "./plans-2018-purecs";
 import { biochemMajor } from "../majors/BIOCHEMISTRY, BS";
-import { Schedule } from "../models/types";
+import { Schedule } from "graduate-common";
 import { mathMajor } from "../majors/MATHEMATICS, BS";
 import { csMajor } from "../majors/COMPUTER SCIENCE, BSCS";
