+++ frontend/src/plans/plans-2018-math.ts
@@ -1,4 +1,4 @@
-import { Schedule } from "../models/types";
+import { Schedule } from "graduate-common";
 
 export const mathPlans: Schedule[] = [
   {
