+++ frontend/src/state/index.ts
@@ -1,12 +1,11 @@
 import { AppState } from "./reducers/state";
 import {
   DNDSchedule,
-  Major,
   CourseWarning,
   IWarning,
   DNDScheduleTerm,
-  Schedule,
 } from "../models/types";
+import { Major, Schedule } from "graduate-common";
 
 /**
  * Utility functions to help extract data from the AppState
