+++ graduate-common/src/index.js
@@ -1,0 +1,5 @@
+import {addPrereqsToSchedules} from "./prereq_loader"
+
+export {
+    addPrereqsToSchedules
+}
